// src/context/types.ts
type ProviderType = "email" | "oauth" | "oidc" | "webauthn";
// ...
export interface SSE_User {
  id?: string;
  name?: string | null;
  email?: string | null;
  image?: string | null;
}

export interface SSE_Profile {
  id?: string | null;
  sub?: string | null;
  name?: string | null;
  given_name?: string | null;
  family_name?: string | null;
  middle_name?: string | null;
  nickname?: string | null;
  preferred_username?: string | null;
  profile?: string | null;
  picture?: string | null | any;
  website?: string | null;
  email?: string | null;
  email_verified?: boolean | null;
  gender?: string | null;
  birthdate?: string | null;
  zoneinfo?: string | null;
  locale?: string | null;
  phone_number?: string | null;
  updated_at?: Date | string | number | null;
  address?: {
    formatted?: string | null;
    street_address?: string | null;
    locality?: string | null;
    region?: string | null;
    postal_code?: string | null;
    country?: string | null;
  } | null;
  [claim: string]: unknown;
}

export interface SSE_Account extends Partial<TokenEndpointResponse> {
  /** Provider's id for this account. E.g. "google". See the full list at https://authjs.dev/reference/core/providers */
  provider: string;
  /**
   * This value depends on the type of the provider being used to create the account.
   * - oauth/oidc: The OAuth account's id, returned from the `profile()` callback.
   * - email: The user's email address.
   * - credentials: `id` returned from the `authorize()` callback
   */
  providerAccountId: string;
  /** Provider's type for this account */
  type: ProviderType;
  /**
   * id of the user this account belongs to
   *
   * @see https://authjs.dev/reference/core/adapters#adapteruser
   */
  userId?: string;
  /**
   * Calculated value based on {@link TokenEndpointResponse.expires_in}.
   *
   * It is the absolute timestamp (in seconds) when the {@link TokenEndpointResponse.access_token} expires.
   *
   * This value can be used for implementing token rotation together with {@link TokenEndpointResponse.refresh_token}.
   *
   * @see https://authjs.dev/guides/refresh-token-rotation#database-strategy
   * @see https://www.rfc-editor.org/rfc/rfc6749#section-5.1
   */
  expires_at?: number;
}

export interface AuthConfig {
// ...
  events?: {
    /**
     * If using a `credentials` type auth, the user is the raw response from your
     * credential provider.
     * For other providers, you'll get the SSE_User object from your adapter, the account,
     * and an indicator if the user was new to your Adapter.
     */
    signIn?: (message: {
      user: SSE_User;
      account?: SSE_Account | null;
      profile?: SSE_Profile;
      isNewUser?: boolean;
    }) => Awaitable<void>;
    /**
     * The message object will contain one of these depending on
     * if you use JWT or database persisted sessions:
     * - `token`: The JWT for this session.
     * - `session`: The session object from your adapter that is being ended.
     */
    signOut?: (
      message:
        | { session: Awaited<ReturnType<Required<Adapter>["deleteSession"]>> }
        | { token: Awaited<ReturnType<JWTOptions["decode"]>> }
    ) => Awaitable<void>;
    createUser?: (message: { user: SSE_User }) => Awaitable<void>;
    updateUser?: (message: { user: SSE_User }) => Awaitable<void>;
    linkAccount?: (message: {
      user: SSE_User | AdapterUser;
      account: SSE_Account;
      profile: SSE_User | AdapterUser;
    }) => Awaitable<void>;
    /**
     * The message object will contain one of these depending on
     * if you use JWT or database persisted sessions:
     * - `token`: The JWT for this session.
     * - `session`: The session object from your adapter.
     */
    session?: (message: { session: Session; token: JWT }) => Awaitable<void>;
  };
// ...
}

/** @internal */
export type InternalProvider<T = ProviderType> = T extends WebAuthnProviderType
  ? WebAuthnConfig
  : T;

export interface InternalOptions<TProviderType = ProviderType> {
  url: URL;
  secret: string | string[];
  theme: Theme;
  debug: boolean;
  provider: InternalProvider<TProviderType>;
  pages: Partial<PagesOptions>;
  session: NonNullable<Required<AuthConfig["session"]>>;
  events: NonNullable<AuthConfig["events"]>;
  jwt: JWTOptions;
  adapter: Required<Adapter> | undefined;
  cookies: Record<keyof CookiesOptions, CookieOption>;
  basePath: string;
  callbacks: NonNullable<Required<AuthConfig["callbacks"]>>;
  callbackUrl: string;
}